refactor(themes): share roving radio nav between both pickers
The /themes page and the in-room ThemeSelector duplicated ~40 lines of
Arrow/Home/End keyboard handling and the retired-theme tab anchor. Extract
useThemeRadioNav; behavior identical (Escape only wired where a popover
exists to close).
