+++ components/ThemeSelector.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { useRef, useCallback } from 'react';
-import { useTheme, visibleThemeIds } from '@/lib/themes';
+import { visibleThemeIds } from '@/lib/themes';
+import { useThemeRadioNav } from '@/hooks/useThemeRadioNav';
 import { ThemeModeControl } from './ThemeModeControl';
 import { ThemePreview } from './ThemePreview';
 import { cn } from '@/lib/utils';
@@ -12,72 +12,13 @@ interface ThemeSelectorProps {
 }
 
 /**
- * Theme picker with self-styling theme rows and unified mode control.
- *
- * Structure:
- * 1. Mode segmented control (Light/Dark/System) at top
- * 2. Vertical list of theme rows
+ * Compact theme picker for the in-room popover (RoomChrome): mode control on
+ * top, self-styling theme rows below. The full specimen experience lives at
+ * /themes; this stays in place so players never navigate away mid-game.
  */
 export function ThemeSelector({ className = '', onClose }: ThemeSelectorProps) {
-  const { themeId, mode, setTheme } = useTheme();
-  const themeListRef = useRef<HTMLDivElement>(null);
-  // A retired active theme (saved before the top-10 roster) renders no row;
-  // anchor the roving tabindex on the first row so keyboard users can enter.
-  const tabAnchorId = visibleThemeIds.includes(themeId)
-    ? themeId
-    : visibleThemeIds[0];
-
-  const handleThemeSelect = (id: string) => {
-    setTheme(id);
-  };
-
-  const focusTheme = useCallback((index: number) => {
-    const buttons = themeListRef.current?.querySelectorAll('[role="radio"]');
-    (buttons?.[index] as HTMLElement)?.focus();
-  }, []);
-
-  const handleThemeKeyDown = useCallback(
-    (event: React.KeyboardEvent) => {
-      const currentIndex = visibleThemeIds.indexOf(themeId);
-
-      switch (event.key) {
-        case 'ArrowDown': {
-          event.preventDefault();
-          const nextIndex = (currentIndex + 1) % visibleThemeIds.length;
-          setTheme(visibleThemeIds[nextIndex]);
-          focusTheme(nextIndex);
-          break;
-        }
-        case 'ArrowUp': {
-          event.preventDefault();
-          const prevIndex =
-            (currentIndex - 1 + visibleThemeIds.length) %
-            visibleThemeIds.length;
-          setTheme(visibleThemeIds[prevIndex]);
-          focusTheme(prevIndex);
-          break;
-        }
-        case 'Home': {
-          event.preventDefault();
-          setTheme(visibleThemeIds[0]);
-          focusTheme(0);
-          break;
-        }
-        case 'End': {
-          event.preventDefault();
-          const lastIndex = visibleThemeIds.length - 1;
-          setTheme(visibleThemeIds[lastIndex]);
-          focusTheme(lastIndex);
-          break;
-        }
-        case 'Escape': {
-          onClose?.();
-          break;
-        }
-      }
-    },
-    [themeId, setTheme, onClose, focusTheme]
-  );
+  const { themeId, mode, setTheme, listRef, tabAnchorId, handleKeyDown } =
+    useThemeRadioNav({ onEscape: onClose });
 
   return (
     <div className={cn('space-y-4', className)}>
@@ -85,11 +26,11 @@ export function ThemeSelector({ className = '', onClose }: ThemeSelectorProps) {
 
       {/* Theme list */}
       <div
-        ref={themeListRef}
+        ref={listRef}
         className="flex flex-col gap-2"
         role="radiogroup"
         aria-label="Select theme"
-        onKeyDown={handleThemeKeyDown}
+        onKeyDown={handleKeyDown}
       >
         {visibleThemeIds.map((id) => (
           <ThemePreview
@@ -97,7 +38,7 @@ export function ThemeSelector({ className = '', onClose }: ThemeSelectorProps) {
             themeId={id}
             isSelected={id === themeId}
             currentMode={mode}
-            onSelect={() => handleThemeSelect(id)}
+            onSelect={() => setTheme(id)}
             tabIndex={id === tabAnchorId ? 0 : -1}
           />
         ))}
